Fire user’s onChange handler when we update the input value internally
We were directly setting `.value` previously. This does not fire the onChange handler given by the user for an input element in React >= 16. However, from a user’s perspective they’re seeing the value change and not being notified which means that their state is out of sync since this works as expected when the user is typing.